added delete event interface

+++ src/server/database/event/event.database.ts
@@ -56,8 +56,13 @@ export const updateEvent = (event: Event, userId: number): Promise<string> => {
     });
 };
 
-export const deleteEvent = (event: Event): Promise<string> => {
+export const deleteEvent = (event: Event, userId: number): Promise<string> => {
     return new Promise<string>(async (resolve, reject) => {
+        const user: User = await readUserById(userId);
+        const userValid: boolean = await validateUserHostingEvent(event, user);
+        if (!userValid) {
+            return reject('User not authorized to edit this event.');
+        }
         // delete all records from UserHostingEvent table
         await database<UserHostingEvent>('UserHostingEvent')
             .delete()
